chore: rename step-section.js to steps-section.js
