+++ components/steps-section.js
